Correctif des composants de la page LifeStyle. Déplacement, refactorisation et correction des imports
