+++ client/src/seo/AvisClientGlobalData.ts
@@ -4,7 +4,7 @@ import AvisClientPortraitDuoControllerData from "../components/PortraitDuoRoot/A
 import AvisClientPortraitFamilleControllerData from "../components/PortraitFamilleRoot/AvisClientPortraitFamilleControllerData";
 import AvisClientPortraitGrossesseControllerData from "../components/PortraitGrossesseRoot/AvisClientPortraitGrossesseControllerData";
 import AvisClientPortraitHobbyControllerData from "../components/PortraitHobbyRoot/AvisClientPortraitHobbyControllerData";
-import AvisClientPortraitLifeStyleControllerData from "../components/PortraitLifeStyleRoot/AvisClientPortraitLifeStyleControllerData";
+import AvisClientPortraitLifeStyleControllerData from "../components/Public/PortraitLifeStyleRoot/AvisClientPortraitLifeStyleControllerData";
 import AvisClientPortraitMariageControllerData from "../components/Public/PortraitMariageRoot/AvisClientPortraitMariageControllerData";
 import AvisClientPortraitSoloControllerData from "../components/Public/PortraitSoloRoot/AvisClientPortraitSoloControllerData";
